Migrate nlearn to IREE Metal backend (trains on M3 GPU)
Runs the model on Apple Silicon via the open-source IREE PJRT stack instead
of the abandoned jax-metal. Verified: full train_step (fwd+bwd+Adam, chunked
CE, attention) compiles and loss decreases on the M3 GPU.

Worktree changes:
- attention.py: detect IREE 'iree_metal' platform (-> standard attention path)
- model.py: COMPUTE_DTYPE fp16 (metal-spirv target lacks bf16)
- train.py: chunked-CE target select via one-hot multiply, not advanced-index
  gather (gather's scatter backward miscompiles under vmap on metal-spirv)
- iree_env.sh: launcher env (patched compiler + lld path + JAX_PLATFORMS)
- iree_risk_probe.py: per-feature backend probes

Plugin + compiler fixes live in ~/src/iree (authored iree_metal_plugin;
SPIRVFinalVectorLowering vector.step patch; abseil arm64 build patch).

+++ model.py
@@ -18,7 +18,11 @@ D_FF        = 2048     # Hidden size of the feed-forward sublayer. Kept at 4 * D
 N_LAYERS    = 4        # How many transformer blocks to stack. GPT-3 has 96.
 MAX_SEQ_LEN = 512      # Maximum context window in tokens. With 3.8x BPE compression, this covers ~2000 characters.
 
-COMPUTE_DTYPE    = jnp.bfloat16  # Forward-pass dtype; stored params stay float32.
+COMPUTE_DTYPE    = jnp.float16  # Forward-pass dtype; stored params stay float32.
+# fp16, not bf16: IREE's metal-spirv target for Apple GPUs advertises fp16 but
+# NOT bf16 (the SPIR-V target capabilities omit it), so bf16 matmuls fail to
+# legalize. fp16 has a narrow exponent range, so training uses loss scaling to
+# keep gradients representable (see train.py).
 
 # ---------------------------------------------------------------------------
 # SECTION 1: EMBEDDINGS
